*: update bazel (#41369)
ref pingcap/tidb#41377
