+++ expression/BUILD.bazel
@@ -175,7 +175,6 @@ go_test(
         "function_traits_test.go",
         "helper_test.go",
         "integration_serial_test.go",
-        "integration_test.go",
         "main_test.go",
         "multi_valued_index_test.go",
         "scalar_function_test.go",
@@ -190,7 +189,6 @@ go_test(
     shard_count = 50,
     deps = [
         "//config",
-        "//domain",
         "//errno",
         "//kv",
         "//parser",
@@ -224,16 +222,12 @@ go_test(
         "//util/mathutil",
         "//util/mock",
         "//util/printer",
-        "//util/sem",
-        "//util/sqlexec",
         "//util/timeutil",
-        "//util/versioninfo",
         "@com_github_gogo_protobuf//proto",
         "@com_github_google_uuid//:uuid",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_pingcap_tipb//go-tipb",
-        "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//oracle",
         "@com_github_tikv_client_go_v2//tikv",
