+++ expression/integration_test/BUILD.bazel
@@ -1,46 +1,43 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_test")
 
 go_test(
-    name = "session_test_test",
-    timeout = "moderate",
+    name = "integration_test_test",
+    timeout = "short",
     srcs = [
+        "integration_test.go",
         "main_test.go",
-        "session_test.go",
-        "temporary_table_test.go",
     ],
     flaky = True,
-    shard_count = 4,
+    shard_count = 50,
     deps = [
         "//config",
         "//domain",
         "//errno",
         "//expression",
         "//kv",
-        "//parser/ast",
         "//parser/auth",
-        "//parser/format",
         "//parser/model",
         "//parser/mysql",
         "//parser/terror",
         "//planner/core",
-        "//privilege/privileges",
         "//session",
-        "//sessionctx",
         "//sessionctx/variable",
-        "//store/copr",
-        "//store/mockstore",
-        "//table/tables",
+        "//table",
+        "//tablecodec",
         "//testkit",
         "//testkit/testmain",
         "//testkit/testsetup",
         "//types",
-        "//util",
-        "//util/memory",
+        "//util/codec",
+        "//util/collate",
+        "//util/sem",
         "//util/sqlexec",
-        "@com_github_pingcap_failpoint//:failpoint",
+        "//util/timeutil",
+        "//util/versioninfo",
+        "@com_github_pingcap_errors//:errors",
+        "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//tikv",
-        "@com_github_tikv_client_go_v2//txnkv/transaction",
         "@org_uber_go_goleak//:goleak",
     ],
 )
