+++ session/sessiontest/BUILD.bazel
@@ -1,0 +1,45 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "sessiontest_test",
+    timeout = "short",
+    srcs = [
+        "main_test.go",
+        "session_test.go",
+    ],
+    flaky = True,
+    shard_count = 50,
+    deps = [
+        "//config",
+        "//domain",
+        "//errno",
+        "//expression",
+        "//kv",
+        "//parser/ast",
+        "//parser/auth",
+        "//parser/format",
+        "//parser/model",
+        "//parser/mysql",
+        "//parser/terror",
+        "//planner/core",
+        "//privilege/privileges",
+        "//session",
+        "//sessionctx",
+        "//sessionctx/variable",
+        "//store/copr",
+        "//store/mockstore",
+        "//table/tables",
+        "//testkit",
+        "//testkit/testmain",
+        "//testkit/testsetup",
+        "//types",
+        "//util",
+        "//util/memory",
+        "//util/sqlexec",
+        "@com_github_pingcap_failpoint//:failpoint",
+        "@com_github_stretchr_testify//require",
+        "@com_github_tikv_client_go_v2//tikv",
+        "@com_github_tikv_client_go_v2//txnkv/transaction",
+        "@org_uber_go_goleak//:goleak",
+    ],
+)
